feat: added forgota and reset password email functionality
