+++ backend/.env
@@ -5,4 +5,4 @@ ALGOR = HS256
 SMTP_SERVER=smtp.gmail.com
 SMTP_PORT=587
 SENDER_EMAIL=josephmaghwe@gmail.com
-SENDER_PASSWORD=ochk daoy xoly kkto
+SENDER_PASSWORD=tctg kkqy nqbt oohu
