+++ backend/app/routes/authenticationRoute.py
@@ -5,6 +5,9 @@ from app.core.database import get_db
 from app.services import authentication
 from fastapi.security import  OAuth2PasswordRequestForm
 from app.schemas.schemaUser import ForgotPasswordRequest, ResetPasswordRequest
+# New imports
+from app.services import authentication
+from app.services import email as email_service 
 
 router = APIRouter(
    
@@ -16,11 +19,12 @@ def login(request: Annotated[OAuth2PasswordRequestForm, Depends()],db:Session= D
     return authentication.login(request,db)
 
 
-
-@router.post('/forgot-password', status_code=status.HTTP_200_OK)
+@router.post("/forgot-password")
 def forgot_password(request: ForgotPasswordRequest, db: Session = Depends(get_db)):
-    return authentication.forgot_password(request, db)
+    # Points to your new email.py service file
+    return email_service.forgot_password(request, db)
 
-@router.post('/reset-password', status_code=status.HTTP_200_OK)
+@router.post("/reset-password")
 def reset_password(request: ResetPasswordRequest, db: Session = Depends(get_db)):
-    return authentication.reset_password(request, db)
+    # Points to your new email.py service file
+    return email_service.reset_password(request, db)
